src/requirements.txt
Flask
twilio
openai
gunicorn
python-dotenv
pymongo
googlesearch-python
requests
beautifulsoup4
